Refactor build job handling: move functions to build_job_handler and update routes
