+++ packages/openci-controller/src/handlers/mod.rs
@@ -1,4 +1,5 @@
 pub mod api_key_handler;
+pub mod build_job_handler;
 pub mod github_repository_handler;
 pub mod github_webhook_handler;
 pub mod user_handler;
